move things around a bit for better structure
